New login page
You need to put the login.html in a  directory called registration. See setting.py changes as well...

Also one of the urls.py is main and the other is specific to the app.

The idea is to have the login come up first and then after a successful login the home page where all the activities happen.